requirements.txt
albumentations==2.0.5
numpy==2.2.3
pandas==2.2.3
Pillow==11.1.0
pynrrd==1.1.3
pytorch_lightning==2.5.0.post0
torch==2.6.0+cu126
torchmetrics==1.6.1
torchvision==0.21.0+cu126
